online_learning/requirements.txt
accessify==0.3.1
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.2.0
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
crashtest==0.4.1
cron-descriptor==1.4.3
cryptography==41.0.7
data==0.4
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
flake8==6.1.0
funcsigs==1.0.2
fuzzywuzzy==0.18.0
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipython==8.18.1
isodate==0.6.1
jaraco.classes==3.3.0
jedi==0.19.1
keyring==24.2.0
kombu==5.3.5
Levenshtein==0.23.0
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
oauth2client==4.1.3
packaging==23.1
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.41
protobuf==4.24.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==7.4.1
pytest-cov==4.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
redis==5.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stripe==8.4.0
thefuzz==0.20.0
tomlkit==0.12.2
traitlets==5.14.0
trove-classifiers==2023.11.9
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.24.6
wcwidth==0.2.12
xattr==0.10.1
zipp==3.17.0